+++ mappers_src/bw_readmapper_src/size_vector.h
@@ -7,11 +7,11 @@
 
 struct size_vector {
     size_t *sizes;
-    int size;
-    int used;
+    size_t size;
+    size_t used;
 };
 
-struct size_vector *empty_size_vector(int initial_size);
+struct size_vector *empty_size_vector(size_t initial_size);
 void delete_size_vector(struct size_vector *v);
 struct size_vector *add_size(struct size_vector *v, size_t size);
