+++ mappers_src/bw_readmapper_src/string_vector.h
@@ -2,13 +2,15 @@
 #ifndef STRING_VECTOR_H
 #define STRING_VECTOR_H
 
+#include <stdlib.h>
+
 struct string_vector {
     char **strings;
-    int size;
-    int used;
+    size_t size;
+    size_t used;
 };
 
-struct string_vector *empty_string_vector(int initial_size);
+struct string_vector *empty_string_vector(size_t initial_size);
 void delete_string_vector(struct string_vector *v);
 
 // when adding a string, we make a copy -- so we know we can
